// src/types/pr.d.ts
/*---------------------------------------------------------------------------------------------
 *  Copyright (c) Simple Tools. All rights reserved.
 *  Licensed under the MIT License. See License.txt in the project root for license information.
 *--------------------------------------------------------------------------------------------*/


export type Pr = {
    title:string,
    policyNumber:string,
    content : {
        text:string;
    }
}